[DROP-IN] Expose option settings for Drop-In UI
