+++ libnavui-dropin/src/test/java/com/mapbox/navigation/dropin/lifecycle/UIComponentTest.kt
@@ -4,6 +4,7 @@ import com.mapbox.navigation.base.ExperimentalPreviewMapboxNavigationAPI
 import com.mapbox.navigation.core.MapboxNavigation
 import com.mapbox.navigation.testing.MainCoroutineRule
 import io.mockk.mockk
+import io.mockk.verify
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.collect
@@ -79,6 +80,25 @@ class UIComponentTest {
 
         assertEquals(listOf(1, 2, 2, 3), component.captured)
     }
+
+    @Test
+    fun `reloadOnChange should re-create and re-attach childComponent on value change`() =
+        runBlockingTest {
+            val child1 = mockk<UIComponent>(relaxed = true)
+            val child2 = mockk<UIComponent>(relaxed = true)
+            val flow = MutableStateFlow(0)
+            val mapboxNavigation = mockk<MapboxNavigation>()
+            val component = reloadOnChange(flow) {
+                if (it % 2 == 0) child1 else child2
+            }
+
+            component.onAttached(mapboxNavigation)
+
+            verify { child1.onAttached(mapboxNavigation) }
+            flow.tryEmit(1)
+            verify { child1.onDetached(mapboxNavigation) }
+            verify { child2.onAttached(mapboxNavigation) }
+        }
 }
 
 class TestUIComponent(initialValue: Int) : UIComponent() {
